+++ package.json
@@ -1,25 +1,75 @@
 {
   "name": "keksobooking",
-  "version": "19.0.0",
+  "version": "20.0.0",
   "private": true,
-  "description": "Личный проект «Кексобукинг» от HTML Academy",
+  "description": "Личный проект «Кексобукинг» от HTML Academy",
+  "devDependencies": {
+    "@htmlacademy/editorconfig-cli": "1.x",
+    "autoprefixer": "9.5.x",
+    "browser-sync": "2.26.x",
+    "del": "^5.1.0",
+    "eslint": "^6.8.0",
+    "gulp": "4.0.x",
+    "gulp-cheerio": "^1.0.0",
+    "gulp-csso": "^4.0.1",
+    "gulp-htmlmin": "^5.0.1",
+    "gulp-imagemin": "^7.1.0",
+    "gulp-plumber": "1.2.x",
+    "gulp-postcss": "8.0.x",
+    "gulp-posthtml": "^3.0.4",
+    "gulp-rename": "^2.0.0",
+    "gulp-sass": "4.0.x",
+    "gulp-sourcemaps": "2.6.x",
+    "gulp-svg-sprite": "^1.5.0",
+    "gulp-svg-sprites": "^4.1.2",
+    "gulp-svgstore": "^7.0.1",
+    "gulp-uglify": "^3.0.2",
+    "gulp-webp": "^4.0.1",
+    "posthtml-include": "^1.4.3",
+    "stylelint": "10.1.x",
+    "stylelint-config-htmlacademy": "0.1.x"
+  },
+  "scripts": {
+    "editorconfig": "editorconfig-cli",
+    "stylelint": "stylelint \"source/sass/**/*.scss\" --syntax scss",
+    "test": "npm run editorconfig && npm run stylelint",
+    "build": "gulp build",
+    "start": "gulp start",
+    "webp": "gulp webp",
+    "images": "gulp images",
+    "compress-js": "gulp compress-js",
+    "refresh": "gulp refresh",
+    "sprite": "gulp sprite",
+    "server": "gulp server"
+  },
+  "browserslist": [
+    "last 2 versions",
+    "IE 11",
+    "Firefox ESR"
+  ],
+  "editorconfig-cli": [
+    "*.json",
+    "*.js",
+    "source/*.html",
+    "source/js/**/*.js",
+    "source/img/**/*.svg",
+    "source/sass/**/*.{sass,scss}"
+  ],
+  "engines": {
+    "node": "8.16"
+  },
+  "dependencies": {
+    "yarn": "^1.22.4"
+  },
+  "main": "gulpfile.js",
   "repository": {
     "type": "git",
-    "url": "git@github.com:htmlacademy-javascript/181350-keksobooking-20.git"
+    "url": "git+https://github.com/Isyumich/181350-keksobooking-20.git"
   },
+  "author": "",
+  "license": "ISC",
   "bugs": {
-    "url": "https://github.com/htmlacademy-javascript/181350-keksobooking-20/issues"
+    "url": "https://github.com/Isyumich/181350-keksobooking-20/issues"
   },
-  "devDependencies": {
-    "eslint": "7.1.0",
-    "eslint-config-htmlacademy": "0.5.2"
-  },
-  "license": "MIT",
-  "dependencies": {},
-  "scripts": {
-    "test": "eslint js/"
-  },
-  "engines": {
-    "node": "8.11"
-  }
+  "homepage": "https://github.com/Isyumich/181350-keksobooking-20#readme"
 }
